+++ CHANGELOG.rst
@@ -2,6 +2,17 @@
 Changelog
 =========
 
+0.3.0 (10-05-2019)
+------------------
+* Explore dashboard
+* Real-time visualization of ExG and orientation signal
+* Device information in Dashboard
+* Environmental data (battery, temperature and light)
+* Real-time bandpass filter
+* New packet structures (ADS1294R & ADS1298R)
+* Heart rate estimation and R-peaks detector in dashboard
+
+
 0.2.0 (2019-03-08)
 ------------------